requirement.txt
Django==4.2.6
qrcode==7.4.2
Pillow==10.0.1
pyzbar==0.1.9
segno==1.5.2


